+++ package.json
@@ -26,11 +26,19 @@
     "@fortawesome/fontawesome-svg-core": "^1.2.32",
     "@fortawesome/pro-light-svg-icons": "^5.15.1",
     "@fortawesome/react-native-fontawesome": "^0.2.6",
+    "@react-native-community/masked-view": "^0.1.10",
+    "@react-navigation/bottom-tabs": "^5.11.2",
+    "@react-navigation/native": "^5.7.3",
+    "@react-navigation/stack": "^5.9.0",
     "@sentry/react-native": "^2.0.2",
     "react": "16.13.1",
     "react-native": "0.63.3",
     "react-native-device-info": "^7.2.1",
     "react-native-dotenv": "^2.4.0",
+    "react-native-gesture-handler": "^1.9.0",
+    "react-native-reanimated": "^1.13.2",
+    "react-native-safe-area-context": "^3.1.9",
+    "react-native-screens": "^2.16.1",
     "react-native-svg": "^12.1.0"
   },
   "devDependencies": {
