+++ src/screens/ExampleScreen.jsx
@@ -1,0 +1,123 @@
+// Third-party dependencies
+import React from 'react'
+import {
+    Button,
+    SafeAreaView,
+    ScrollView,
+    StyleSheet,
+    StatusBar,
+    Text,
+    View,
+} from 'react-native'
+import {
+    Colors,
+    DebugInstructions,
+    Header,
+    LearnMoreLinks,
+    ReloadInstructions,
+} from 'react-native/Libraries/NewAppScreen'
+import {
+    FontAwesomeIcon
+} from '@fortawesome/react-native-fontawesome'
+import {
+    faUfoBeam
+} from '@fortawesome/pro-light-svg-icons'
+import {
+    useNavigation
+} from "@react-navigation/native"
+
+function ExampleScreen() {
+    const navigation = useNavigation()
+
+    return (
+        <>
+            <StatusBar barStyle="dark-content" />
+            <SafeAreaView>
+                <ScrollView
+                    contentInsetAdjustmentBehavior="automatic"
+                    style={styles.scrollView}>
+                    <Header />
+                    {global.HermesInternal == null ? null : (
+                        <View style={styles.engine}>
+                            <Text style={styles.footer}>Engine: Hermes</Text>
+                        </View>
+                    )}
+                    <View style={styles.body}>
+                        <View style={styles.sectionContainer}>
+                            <Text style={styles.sectionTitle}>Step One</Text>
+                            <Text style={styles.sectionDescription}>
+                                Edit <Text style={styles.highlight}>App.js</Text> to change this
+                                screen and then come back to see your edits.
+                            </Text>
+                            <FontAwesomeIcon icon={faUfoBeam} size={50} />
+                            <Button
+                                title="Go to OtherScreen"
+                                onPress={() => navigation.navigate('ExampleScreen2')}
+                            />
+                        </View>
+                        <View style={styles.sectionContainer}>
+                            <Text style={styles.sectionTitle}>See Your Changes</Text>
+                            <Text style={styles.sectionDescription}>
+                                <ReloadInstructions />
+                            </Text>
+                        </View>
+                        <View style={styles.sectionContainer}>
+                            <Text style={styles.sectionTitle}>Debug</Text>
+                            <Text style={styles.sectionDescription}>
+                                <DebugInstructions />
+                            </Text>
+                        </View>
+                        <View style={styles.sectionContainer}>
+                            <Text style={styles.sectionTitle}>Learn More</Text>
+                            <Text style={styles.sectionDescription}>
+                                Read the docs to discover what to do next:
+                            </Text>
+                        </View>
+                        <LearnMoreLinks />
+                    </View>
+                </ScrollView>
+            </SafeAreaView>
+        </>
+    )
+}
+
+const styles = StyleSheet.create({
+    scrollView: {
+        backgroundColor: Colors.lighter,
+    },
+    engine: {
+        position: 'absolute',
+        right: 0,
+    },
+    body: {
+        backgroundColor: Colors.white,
+    },
+    sectionContainer: {
+        marginTop: 32,
+        paddingHorizontal: 24,
+    },
+    sectionTitle: {
+        fontSize: 24,
+        fontWeight: '600',
+        color: Colors.black,
+    },
+    sectionDescription: {
+        marginTop: 8,
+        fontSize: 18,
+        fontWeight: '400',
+        color: Colors.dark,
+    },
+    highlight: {
+        fontWeight: '700',
+    },
+    footer: {
+        color: Colors.dark,
+        fontSize: 12,
+        fontWeight: '600',
+        padding: 4,
+        paddingRight: 12,
+        textAlign: 'right',
+    },
+})
+
+export default ExampleScreen
